:sparkles: clean up code, rewrite README
